Fixed ticket security issues and fixed auth fr

+++ app/Http/Controllers/TicketController.php
@@ -3,12 +3,10 @@
 namespace App\Http\Controllers;
 
 use App\Models\Ticket;
-use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Log;
 use App\Services\TicketService;
 use App\Http\Requests\CreateTicketRequest;
 use App\Http\Requests\HandleTicketRequest;
-use App\Http\Requests\RequestedDataRequest;
 
 class TicketController extends Controller
 {
@@ -54,13 +52,15 @@ class TicketController extends Controller
 
     public function createTicket(CreateTicketRequest $request)
     {
+        $user = auth()->user();
+
+        if ( ! $user) {
+            return response()->json(['message' => 'Usuário não autorizado'], 403);
+        }
+
         try {
-            $message = $this->ticketService->createTicket($request->validated());
-            return response()->json(['message' => $message], 200);
-        } catch (\Illuminate\Validation\ValidationException $e) {
-            return response()->json(['errors' => 'Input inválido'], 422);
-        } catch (\InvalidArgumentException $e) {
-            return response()->json(['error' => $e->getMessage()], 400);
+            $this->ticketService->createTicket($request->validated(), $user);
+            return response()->json(['message' => 'Ticket criado com sucesso'], 200);
         } catch (\Exception $e) {
             Log::error($e);
             return response()->json(['message' => 'Ocorreu um erro'], 500);
@@ -94,6 +94,10 @@ class TicketController extends Controller
     //DEBUG
     public function dropAllTickets()
     {
+        if ( ! env('APP_DEBUG')) {
+            return response()->json(['message' => 'Essa rota só pode ser acessada em ambiente de desenvolvimento'], 403);
+        }
+
         try {
             Ticket::truncate();
             return response()->json(['message' => 'Tickets deletados com sucesso'], 200);
